// kodikosERP-master/src/app/models/formatki/assortmentform.model.ts
export interface Assortmentform {
  comments?: boolean;
  status?: boolean;
  buttonDodajKontr?: boolean;
  buttonOfertaKontr?: boolean;
  buttonEdytujKontr?: boolean;
  buttonUsunKontr?: boolean;
  projekty?: boolean;
  firmy?: boolean;
  widziWszystkie?: boolean;
  pliki?: boolean;
  buttonPliki?: boolean;
}